refactor: setup values and get initial syn lst
